Add files via upload
Hey I think I've managed to combine the evap into the time_evolve_sorting function and renamed it time_evolve_sorting_and_evap! ;) My most creative naming moment! I have mainly added some more print statements and done them so that it is more easily understandable from the what the kernel prints what is going on. I have added the total number of atoms currently in the system to x-y graph just to check that we are actually evapping atoms. Hopefully if you run it as it is you will see what I mean. It shouldn't take longer than a minute, only because number of atoms and cells are low :( . In summary I actually haven't done much really just put in the evap and then made the print statements more appealing to read :)